test: move internal/database/test/postgres.go to internal/database/test/runtime_postgres package
